+++ src/client.ts
@@ -303,13 +303,10 @@ export class SubscriptionClient {
 
           break;
         case SUBSCRIPTION_DATA:
-          if (parsedMessage.payload.data && !parsedMessage.payload.errors) {
-              this.subscriptions[subId].handler(null, parsedMessage.payload.data);
-          } else {
-            this.subscriptions[subId].handler(this.formatErrors(parsedMessage.payload.errors), null);
-          }
+          const payloadData = parsedMessage.payload.data || null;
+          const payloadErrors = parsedMessage.payload.errors ? this.formatErrors(parsedMessage.payload.errors) : null;
+          this.subscriptions[subId].handler(payloadErrors, payloadData);
           break;
-
         case KEEPALIVE:
           break;